GER-338: Installer changes for 2.11.3
Repository configuration must be updated to make sure
denyNonFastForwards is set to false on every Gerrit
managed GitMS repository. The installer will now detect
replicated repositories where this setting is not present
and offer to set them in a replicated method.

Change-Id: Ia0a9c392abf3f39841f80075bbf1824808d52ed8

GER-319 + GER-345:

Fix small errors in installer.sh
Remove slave-wait tables from install
Enable NFF updates on repos in an upgrade

Change-Id: I0d3e515b3cb8c8244794e6dfeba74e64e8ad0f69

Fix syntax error

Some fixes for the initial stress tests for Gerrit

- The queue was getting full, due to not polling her enough!
- The db was still reporting deadlocks, but that is Chris' fault! (did
  not include my new code in the first installer)
- More messages in the log files to understand if things go wrong

Change-Id: I875b5f347c9513098233a6e97822708f825a8c5b

Update JGit dependency

Testing a new version of JGit that has replaced lastModified
calls with their NIO equilvalent

Change-Id: I7afad86bc55acddd934cc083687bc3ceddd3bd7c

GER-260: Error handling changes for Gerrit UI

Changed how the UI displays errors caused by the replicator when
creating/deleting a branch.

Change-Id: Ifd175d1cb78a1a4f9896bfbe897ee1db97e0b43f

Removed call to add_master_slave_db_tables

Change-Id: I6facd89cae9056af4fded2ff8b39f04994606b7c

GER-261: Better feedback on repository deployment

Improved error messages propagated to UI when repo
creation fails in Gerrit.

Change-Id: Ib29a552f6d86cab8764a69306d9fbce2c5794f05

GER-359, GER-356, GER-350, GER-268, GER-357

Still testing, and still to remove some log.info which are there to
check everything is ok. Will be removed shortly.

This, in conjunction with the fix on the gerrit integration tests, fixes
the problem about the abnormal number of index changes events we were
seeing on the dger machines.
Introduces new threads to manage the in and out of index events.
On the landing node, if the index is not yet ready on the database, wait
max for an hour to reindex that change. As soon as it sees the date is
ok between the change itself and the one on the db, it allows the
reindex. If it's not yet the time, it pushes back the change in the
queue.

There is a queue which stores the changes coming from Gerrit. This
queue is then read and filtered to avoid duplicates and then sent to the
other nodes.
On the receiving node the changes are stored as files (as part of a
replicated stream event) and then reindexed.

Change-Id: I516d01abe11c82b7a3e47cfc9dfbf46c78eedd73

Code reintroduced after mistaken removal

Change-Id: I5bc7e7b88dbc466e1aea4d08f36d71b9f262e347
(cherry picked from commit db27d6a85334b7fd88428d86e3370f6ff6fab6ac)